uniquely handle the supplemental rows

+++ app/scraper/contests.py
@@ -28,9 +28,11 @@ def scrape_contests():
     deleted_count = 0
     parsed_count = 0
     supplemented_count = 0
+    group_count = 0
 
     for group in sources[election]:
         source = sources[election][group]
+        group_count = group_count + 1
 
         if 'type' in source and source['type'] == 'results':
             # handle parsed contests
@@ -47,25 +49,25 @@ def scrape_contests():
             # commit parsed rows
             db.session.commit()
             
-            # Handle post processing actions
-            supplemental = contest.post_processing('contests')
-            for supplemental_contest in supplemental:
-                rows = supplemental_contest['rows']
-                action = supplemental_contest['action']
-                if action is not None:
+    # Handle post processing actions. this only needs to happen once, not for every group.
+    supplemental = contest.post_processing('contests')
+    for supplemental_contest in supplemental:
+        rows = supplemental_contest['rows']
+        action = supplemental_contest['action']
+        if action is not None and rows != []:
+            for row in rows:
+                if row is not []:
                     if action == 'insert' or action == 'update':
-                        for row in rows:
-                            db.session.merge(row)
-                            if action == 'insert':
-                                inserted_count = inserted_count + 1
-                            elif action == 'update':
-                                updated_count = updated_count + 1
+                        db.session.merge(row)
+                        if action == 'insert':
+                            inserted_count = inserted_count + 1
+                        elif action == 'update':
+                            updated_count = updated_count + 1
+                        supplemented_count = supplemented_count + 1
                     elif action == 'delete':
-                        for row in rows:
-                            db.session.delete(row)
-                            deleted_count = deleted_count + 1
-                    supplemented_count = supplemented_count + 1
-            # commit supplemental rows
-            db.session.commit()
+                        db.session.delete(row)
+                        deleted_count = deleted_count + 1
+    # commit supplemental rows
+    db.session.commit()
 
-    return "Rows inserted: %s; Rows updated: %s; Rows deleted: %s. Parsed rows: %s Supplemental rows: %s" % (str(inserted_count), str(updated_count), str(deleted_count), str(parsed_count), str(supplemented_count))
+    return "Elections scanned: %s. Rows inserted: %s; Rows updated: %s; Rows deleted: %s. Parsed rows: %s Supplemental rows: %s" % (str(group_count), str(inserted_count), str(updated_count), str(deleted_count), str(parsed_count), str(supplemented_count))
